tirando o codigo do PUT user
